Employees now complete dishes
Employees now complete dishes relative to how much they make. Had to switch dish count from int to long for calculations, and to account for the fact that int only goes up to 2.1 billion.

+++ Assets/DishSpawner.cs
@@ -22,7 +22,7 @@ public class DishSpawner : MonoBehaviour
         }
     }
 
-    public DishData GetRandomDish(int totalDishesCleaned)
+    public DishData GetRandomDish(long totalDishesCleaned)
     {
         // Unlock new dishes when threshold is reached
         foreach (var dish in allDishes)
